feat: home page displays student real info
Now, the values ​​scraped from CAGR are used for the creation of the home page. Another important factor of this commit is that the project was restructured, no longer following the MVC standard, despite maintaining object-oriented programming. I believe that these changes will be beneficial for the project since they make the development with kivy much more fluid, although refactoring will be necessary in the next commits.